updated test scripts

+++ toolbox/anomaly/pipelines/utils.py
@@ -4,4 +4,4 @@ from dataclasses import dataclass
 @dataclass
 class Reconstruction():
     timestamp: pd.Timestamp = None 
-    reconstruction_error: int = None
+    reconstruction_error: float = None
